for #2567, move Condition and AndCondition to optimize module
