+++ sharding-core/sharding-core-optimize/src/test/java/org/apache/shardingsphere/core/optimize/statement/encrypt/condition/PredicateUtilsTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.parse.filler.impl.dml;
+package org.apache.shardingsphere.core.optimize.statement.encrypt.condition;
 
 import org.apache.shardingsphere.core.parse.sql.segment.dml.expr.ExpressionSegment;
 import org.apache.shardingsphere.core.parse.sql.segment.dml.expr.complex.SubquerySegment;
